@@ -80,8 +80,8 @@
         "drupal/core-composer-scaffold": "^10.3",
         "drupal/core-recommended": "^10.3",
         "drupal/csv_serialization": "^4.0",
-        "drupal/date_recur": "^3.2",
-        "drupal/date_recur_modular": "^3.0",
+        "drupal/date_recur": "^3.8",
+        "drupal/date_recur_modular": "^3.2",
         "drupal/dependent_fields": "^1.0",
         "drupal/disable_html5_validation": "^1.1",
         "drupal/dropzonejs": "^2.8",
@@ -209,7 +209,7 @@
     "prefer-stable": true,
     "config": {
         "platform": {
-            "php": "8.2"
+            "php": "8.3"
         },
         "discard-changes": true,
         "sort-packages": true,
